scraper/requirements.txt
requests
beautifulsoup4
SQLAlchemy>=1.4
pyodbc
python-dotenv
aiohttp
